requirements.txt
anytree==2.6.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
delegator.py==0.1.1
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==3.0.2
pexpect==4.6.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.4.0
python-dotenv==0.10.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
timeout-decorator==0.4.1
uritemplate==3.0.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
